+++ core/chains/evm/config/config_test.go
@@ -277,7 +277,7 @@ func TestChainScopedConfig_GasEstimator(t *testing.T) {
 	assert.Equal(t, assets.GWei(20), ge.PriceDefault())
 	assert.Equal(t, assets.GWei(500), ge.PriceMax())
 	assert.Equal(t, assets.GWei(1), ge.PriceMin())
-	assert.Equal(t, uint32(500000), ge.LimitDefault())
+	assert.Equal(t, uint64(500000), ge.LimitDefault())
 	assert.Equal(t, uint32(500000), ge.LimitMax())
 	assert.Equal(t, float32(1), ge.LimitMultiplier())
 	assert.Equal(t, uint32(21000), ge.LimitTransfer())
@@ -317,7 +317,7 @@ func TestChainScopedConfig_Profiles(t *testing.T) {
 	tests := []struct {
 		name                           string
 		chainID                        int64
-		expectedGasLimitDefault        uint32
+		expectedGasLimitDefault        uint64
 		expectedMinimumContractPayment string
 	}{
 		{"default", 0, 500000, "0.00001"},
