+++ core/services/fluxmonitorv2/mocks/orm.go
@@ -48,7 +48,7 @@ func (_m *ORM) CountFluxMonitorRoundStats() (int, error) {
 }
 
 // CreateEthTransaction provides a mock function with given fields: ctx, fromAddress, toAddress, payload, gasLimit, idempotencyKey
-func (_m *ORM) CreateEthTransaction(ctx context.Context, fromAddress common.Address, toAddress common.Address, payload []byte, gasLimit uint32, idempotencyKey *string) error {
+func (_m *ORM) CreateEthTransaction(ctx context.Context, fromAddress common.Address, toAddress common.Address, payload []byte, gasLimit uint64, idempotencyKey *string) error {
 	ret := _m.Called(ctx, fromAddress, toAddress, payload, gasLimit, idempotencyKey)
 
 	if len(ret) == 0 {
@@ -56,7 +56,7 @@ func (_m *ORM) CreateEthTransaction(ctx context.Context, fromAddress common.Addr
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, common.Address, common.Address, []byte, uint32, *string) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, common.Address, common.Address, []byte, uint64, *string) error); ok {
 		r0 = rf(ctx, fromAddress, toAddress, payload, gasLimit, idempotencyKey)
 	} else {
 		r0 = ret.Error(0)
