+++ core/services/ocrcommon/transmitter.go
@@ -27,7 +27,7 @@ type Transmitter interface {
 type transmitter struct {
 	txm                         txManager
 	fromAddresses               []common.Address
-	gasLimit                    uint32
+	gasLimit                    uint64
 	effectiveTransmitterAddress common.Address
 	strategy                    types.TxStrategy
 	checker                     txmgr.TransmitCheckerSpec
@@ -39,7 +39,7 @@ type transmitter struct {
 func NewTransmitter(
 	txm txManager,
 	fromAddresses []common.Address,
-	gasLimit uint32,
+	gasLimit uint64,
 	effectiveTransmitterAddress common.Address,
 	strategy types.TxStrategy,
 	checker txmgr.TransmitCheckerSpec,
