+++ core/services/pipeline/task.eth_call_test.go
@@ -33,7 +33,7 @@ func TestETHCallTask(t *testing.T) {
 	testutils.SkipShortDB(t)
 
 	var specGasLimit uint32 = 123
-	const gasLimit uint32 = 500_000
+	const gasLimit uint64 = 500_000
 	const drJobTypeGasLimit uint32 = 789
 
 	tests := []struct {
